Bookbuilder modify keeps orders (#174)
* BookBuilderTest.ts: add asserts on adding and modifying orders with zero size.

* BookBuilder#modify: remove redundant check for order existence.

Because BookBuilder#remove() first checks if the order exists and
returns null if it doesn't, there is no need for modify() to also
check for its existence.

* BookBuilder#modify: keep an order even if its size goes to zero.

This is to work around a bug where Trader#handleTradeExecutedMessage()
would modify the order size to zero which would cause it to be deleted
from the orderbook. Later when Trader#handleTradeFinalized() was
called it wouldn't find the order and wouldn't emit
Trader.trade-finalized. Now with d3103dd committed it will always emit
Trader.trade-finalized.

However, this still appears to be the correct action. BookBuilder
supports adding orders with 0 size, which is inconsistent with
modifying an error. Also, only deleting orders when the trade is
finalized seems like a good consistency thing.

* Trader#handleTradeFinalized: micro-optimize.